+++ src/sdc11073/consumer/consumerimpl.py
@@ -298,6 +298,12 @@ class SdcConsumer:
             self.msg_reader, self.msg_factory, self._logger, self._services_dispatcher)
         self._alternative_hostname = alternative_hostname
 
+        # parameters of start_all call, will be set later in start_all
+        self._not_subscribed_actions_param: Iterable[str] | None = None
+        self._fixed_renew_interval_param: float | None = None
+        self._shared_http_server_param: Any | None = None
+        self._check_get_service_param: bool | None = None
+
     def set_mdib(self, mdib: ConsumerMdib | None):
         """SdcConsumer sometimes must know the mdib data (e.g. Set service, activate method)."""
         if mdib is not None and self._mdib is not None:
@@ -478,6 +484,10 @@ class SdcConsumer:
                which is the minimal requirement for a sdc provider.
         :return: None
         """
+        self._not_subscribed_actions_param = not_subscribed_actions
+        self._fixed_renew_interval_param = fixed_renew_interval
+        self._shared_http_server_param = shared_http_server
+        self._check_get_service_param = check_get_service
         self._logger.debug('connecting to %s', self._device_location)
         self._connect()
         self._logger.debug('reading meta data from %s', self._device_location)
@@ -573,6 +583,17 @@ class SdcConsumer:
             client.close()
         self._soap_clients = {}
         self._stop_event_sink()
+
+    def restart(self):
+        """forget existing data and restart from the beginning."""
+        mdib = self._mdib  # keep existing mdib connection
+        self.stop_all()  # with unsubscribe
+        # start with the same parameters as initially
+        self.start_all(self._not_subscribed_actions_param,
+                       self._fixed_renew_interval_param,
+                       self._shared_http_server_param,
+                       self._check_get_service_param)
+        self.set_mdib(mdib)
 
     def set_used_compression(self, *compression_methods: str):
         """Use only one of these compression methods."""